A PoC running security tools in a standard way
The idea is to provide an adaptor for tools so that the
tools can be run seamless together just by adding their names.

This might allow for opinionated integrations that 'just work' for users
or maybe it will all end in tears.

Lots more todo before this is useful.

NOTE: It's been a while since I've worked with docker...

Signed-off-by: Tom Hennen <tom.hennen@gmail.com>